Reorganized Modules, all 3 types have query and compression functions
